# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API sobre el Sistema de Transporte Colectivo (STC) de la Ciudad de México",
        "title": "Apimetro",
        "contact": {
            "name": "galigaribaldi (Galileo Cabrera Garibaldi)",
            "email": "galigaribaldi0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/movilidad",
    "paths": {
        "/mapas/geojsonEstacion": {
            "get": {
                "description": "Obtener estaciones en formato GeoJSON con filtros avanzados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeoJSON"
                ],
                "summary": "GeoJSON Estaciones",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by sistema",
                        "name": "sistema",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by num_comercial",
                        "name": "num_comercial",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by alcaldia_municipio",
                        "name": "alcaldia_municipio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by nombre_ramal",
                        "name": "nombre_ramal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by jerarquia_transporte",
                        "name": "jerarquia_transporte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by derecho_de_via",
                        "name": "derecho_de_via",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by es_cetram",
                        "name": "es_cetram",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by nombre_cetram",
                        "name": "nombre_cetram",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by cetram_real (250m radius)",
                        "name": "cetram_real",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FeatureCollection"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/{sistema}/estacion": {
            "get": {
                "description": "Obtener datos a través de filtros como nombre, línea, municipio, colores y CETRAM",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estacion"
                ],
                "summary": "Datos de Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID interno de estación",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre de la estación",
                        "name": "nombre",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID de la línea",
                        "name": "linea_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Alcaldía o municipio",
                        "name": "alcaldia_municipio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Color de línea en español",
                        "name": "color_esp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Color de línea en inglés",
                        "name": "color_en",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Año de inauguración",
                        "name": "anio",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtra estaciones CETRAM",
                        "name": "es_cetram",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Estacion"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Crear una nueva estación en el sistema de transporte",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estacion"
                ],
                "summary": "Crear Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la estación a crear",
                        "name": "estacion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Estacion"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Eliminar una estación por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estacion"
                ],
                "summary": "Eliminar Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la estación a eliminar",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualizar los datos de una estación por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Estacion"
                ],
                "summary": "Actualizar Estacion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sistema de transporte",
                        "name": "sistema",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID de la estación a actualizar",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la estación",
                        "name": "estacion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Estacion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
        "models.Estacion": {
            "type": "object",
            "properties": {
                "alcaldia_municipio": {
                    "type": "string",
                    "example": "Miguel Hidalgo"
                },
                "anio": {
                    "type": "string",
                    "example": "1980"
                },
                "cve_est": {
                    "type": "string",
                    "example": "TAC"
                },
                "es_cetram": {
                    "type": "boolean",
                    "example": false
                },
                "estacion_id": {
                    "type": "integer",
                    "example": 1
                },
                "estacion_id_oficial": {
                    "type": "integer",
                    "example": 1001
                },
                "estado_ciudad": {
                    "type": "string",
                    "example": "Ciudad de México"
                },
                "existe": {
                    "type": "boolean",
                    "example": true
                },
                "latitud": {
                    "type": "number",
                    "example": 19.403
                },
                "linea_id": {
                    "type": "integer",
                    "example": 1
                },
                "longitud": {
                    "type": "number",
                    "example": -99.188
                },
                "nombre": {
                    "type": "string",
                    "example": "Tacubaya"
                },
                "nombre_cetram": {
                    "type": "string",
                    "example": "Tacubaya CETRAM"
                },
                "num_estacion": {
                    "type": "integer",
                    "example": 1
                },
                "sistema": {
                    "type": "string",
                    "example": "METRO"
                },
                "stop_gtfs": {
                    "type": "string",
                    "example": "STC_001"
                },
                "tipo": {
                    "type": "string",
                    "example": "Terminal"
                }
            }
        },
        "models.Feature": {
            "type": "object",
            "properties": {
                "geometry": {
                    "description": "Geometría de tipo Point",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Geometry"
                        }
                    ]
                },
                "properties": {
                    "description": "Atributos dinámicos",
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "description": "\"Feature\"",
                    "type": "string",
                    "example": "Feature"
                }
            }
        },
        "models.FeatureCollection": {
            "type": "object",
            "properties": {
                "features": {
                    "description": "Lista de features (líneas o estaciones)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Feature"
                    }
                },
                "type": {
                    "description": "\"FeatureCollection\"",
                    "type": "string",
                    "example": "FeatureCollection"
                }
            }
        },
        "models.Geometry": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "description": "interface{} permite múltiples niveles de arreglos"
                },
                "type": {
                    "description": "\"Point\"",
                    "type": "string",
                    "example": "Point"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}